feat(ways): split lint.rs + x- escape + --fix removal + locale stub validation
ways lint had grown to 938 lines in one file — right at the quality
way's "priority split" threshold. This commit splits it along its
natural seams and, in the same pass, closes three long-standing gaps
in schema-driven validation.

Split (functional refactor, no behavior change on its own):
- cmd/lint/mod.rs — `run()` entry point, module declarations
- cmd/lint/schema.rs — Schema struct + loaders + is_reserved_field
- cmd/lint/helpers.rs — frontmatter string primitives (extract, parse,
  has_field, get_field_value, when_block, multi-line collapse) plus
  two new removal helpers (remove_top_level_field,
  remove_when_subfield) with unit tests
- cmd/lint/per_file.rs — all the per-file lint rules
- cmd/lint/scanning.rs — directory walker
- cmd/lint/locale_stubs.rs — new locale-stub validator pass
- cmd/lint/requires.rs — ADR-116 macro-scan + shell tokenizer
- cmd/lint/provenance.rs — ADR-110 sidecar validator

Features (what the user wanted):

1. Foreign fields are now actually removable. UNKNOWN warnings have
   always fired on fields not in the schema, but --fix ignored them.
   Now --fix calls remove_top_level_field / remove_when_subfield and
   emits `FIXED: <path> — removed foreign field 'xxx'`. The top-level
   removal preserves any indented continuation (so removing a block-
   shaped foreign field cleans out its nested keys too). The
   continuation detection stops at the next top-level key so adjacent
   fields aren't eaten.

2. x-* prefix escape hatch (is_reserved_field in schema.rs). Follows
   the JSON Schema / OpenAPI / package.json convention — fields whose
   name starts with `x-` are intentionally foreign and the linter
   leaves them alone. Use cases: prototyping a schema extension,
   marking a "mostly correct" misspelling so --fix doesn't eat it,
   carrying metadata the schema doesn't know about. Applies to way
   frontmatter, when: sub-fields, and locale stub keys uniformly.

3. Locale stubs (*.locales.jsonl) are now validated. The previous
   lint scanner filtered to .md only, so per-language overrides got
   zero schema checking. New `lint/locale_stubs.rs` walks the scan dir
   for *.locales.jsonl files, validates each line against a new
   `locale_stub:` section in frontmatter-schema.yaml listing the four
   valid keys (lang, description, vocabulary, embed_threshold).
   Unknown keys → UNKNOWN warning; --fix removes them via
   serde_json::Map::shift_remove (NOT remove, which defaults to
   swap_remove even with preserve_order — that would churn git diffs
   by swapping the tail into the removed slot).

Supporting change: serde_json gains the `preserve_order` feature in
ways-cli/Cargo.toml. Required for shift_remove and for reading stubs
in their original key order. Does not affect any other serde_json
usage in the crate.

Schema: frontmatter-schema.yaml gains a `locale_stub:` top-level block
listing the four valid keys with type and role. Keeps the "schema is
the single source of truth" property the main validator already has.

Output format: `ways lint` now prints "scanned N way files, M locale
stub files" in its summary line. On a clean repo: 100 way files + 83
locale stubs.

Tested end-to-end by injecting a foreign top-level field plus an
x-experimental escape-hatch field into a real way file, and a
bogus_stub_key + x-experimental into a real locale stub line:
- Without --fix: two UNKNOWN warnings (not four — x- fields skipped)
- With --fix: two FIXED lines, bogus entries removed in-place,
  x-experimental preserved at its original position, key order of
  surrounding fields unchanged.
