+++ tools/ways-cli/Cargo.toml
@@ -17,7 +17,7 @@ tui = ["crossterm"]
 clap = { version = "4", features = ["derive"] }
 serde = { version = "1", features = ["derive"] }
 serde_yaml = "0.9"
-serde_json = "1"
+serde_json = { version = "1", features = ["preserve_order"] }
 walkdir = "2"
 anyhow = "1"
 glob = "0.3"
